Fix architectures, improve ci
