+++ .github/workflows/docker-images.yml
@@ -6,7 +6,39 @@ on:
       - 'v*'
 
 jobs:
-  build:
+  build-auth:
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: Check out the repo
+        uses: actions/checkout@v4
+
+      - name: Log in to Docker Hub
+        uses: docker/login-action@v3
+        with:
+          username: ${{ secrets.DOCKERHUB_USERNAME }}
+          password: ${{ secrets.DOCKERHUB_TOKEN }}
+
+      - name: Extract metadata (tags, labels) for Docker
+        id: meta
+        uses: docker/metadata-action@v5
+        with:
+          images: epigrapho/gaalafis-lfs-auth
+          tags: |
+            type=semver,pattern={{version}}
+            type=semver,pattern={{major}}.{{minor}}
+            type=sha
+
+      - name: Build the Docker image
+        uses: docker/build-push-action@v2
+        with:
+          context: ./modules/auth
+          file: ./modules/auth/Dockerfile
+          tags: ${{ steps.meta.outputs.tags }}
+          labels: ${{ steps.meta.outputs.labels }}
+          push: true
+
+  build-lfs-server:
     runs-on: ubuntu-latest
 
     strategy:
