--- modules/auth/.env
@@ -1,2 +1,0 @@
-BASE_URL=https://proxy/
-JWT_SECRET_FILE=/run/secrets/jwt_secret
