+++ modules/lfs-server/Dockerfile
@@ -11,9 +11,11 @@ FROM chef AS builder
 COPY --from=planner /app/recipe.json recipe.json
 RUN cargo chef cook --release --recipe-path recipe.json
 COPY . .
-RUN cargo build --release --bin lfs-info-server
-RUN cargo build --release --bin sbs_custom_signing
-RUN cargo build --release --bin sbs_custom_signing_locks
+
+RUN cargo build --release --bin server-proxy-sbs
+RUN cargo build --release --bin server-proxy-sbs-locks-pg
+RUN cargo build --release --bin server-signer-sbs
+RUN cargo build --release --bin server-signer-sbs-locks-pg
 
 FROM ubuntu:22.04 AS runtime
 WORKDIR /app
@@ -21,14 +23,18 @@ RUN apt-get update && apt-get install -y openssl
 EXPOSE 3000
 
 FROM runtime AS runtime-proxy-sbs
-COPY --from=builder /app/target/release/sbs_custom_signing /usr/local/bin/server
+COPY --from=builder /app/target/release/server-proxy-sbs /usr/local/bin/server
 ENTRYPOINT ["/usr/local/bin/server"]
 
 FROM runtime AS runtime-proxy-sbs-locks-pg
-COPY --from=builder /app/target/release/sbs_custom_signing_locks /usr/local/bin/server
+COPY --from=builder /app/target/release/server-proxy-sbs-locks-pg /usr/local/bin/server
+ENTRYPOINT ["/usr/local/bin/server"]
+
+FROM runtime AS runtime-signer-sbs
+COPY --from=builder /app/target/release/server-signer-sbs /usr/local/bin/server
 ENTRYPOINT ["/usr/local/bin/server"]
 
 FROM runtime AS runtime-signer-sbs-locks-pg
-COPY --from=builder /app/target/release/lfs-info-server /usr/local/bin/server
+COPY --from=builder /app/target/release/server-signer-sbs-locks-pg /usr/local/bin/server
 ENTRYPOINT ["/usr/local/bin/server"]
