+++ modules/lfs-server/src/bin/server-signer-sbs-locks-pg.rs
@@ -21,6 +21,7 @@ use lfs_info_server::{
         services::Services,
         token_encoder_decoder::TokenEncoderDecoder,
     },
+    server::RouterExt,
 };
 
 /* -------------------------------------------------------------------------- */
@@ -112,11 +113,11 @@ async fn main() {
     // build our application with a route
     let app = Router::new()
         // `POST /objects/batch?repo=a/b/c`
-        .route("/objects/batch", post(post_objects_batch))
-        .route("/locks", post(post_lock))
-        .route("/locks", get(list_locks))
-        .route("/locks/:id/unlock", post(unlock))
-        .route("/locks/verify", post(list_locks_for_verification))
+        .directory_route("/objects/batch", post(post_objects_batch))
+        .directory_route("/locks", post(post_lock))
+        .directory_route("/locks", get(list_locks))
+        .directory_route("/locks/:id/unlock", post(unlock))
+        .directory_route("/locks/verify", post(list_locks_for_verification))
         // Error handling
         .layer(middleware::from_fn(handle_and_filter_error_details))
         .with_state(services);
