+++ modules/lfs-server/src/server.rs
@@ -1,0 +1,22 @@
+use axum::body::HttpBody;
+use axum::Router;
+use axum::routing::MethodRouter;
+
+pub trait RouterExt<S, B>
+    where
+        B: HttpBody + Send + 'static,
+        S: Clone + Send + Sync + 'static,
+{
+    fn directory_route(self, path: &str, method_router: MethodRouter<S, B>) -> Self;
+}
+
+impl<S, B> RouterExt<S, B> for Router<S, B>
+    where
+        B: HttpBody + Send + 'static,
+        S: Clone + Send + Sync + 'static,
+{
+    fn directory_route(self, path: &str, method_router: MethodRouter<S, B>) -> Self {
+        self.route(path, method_router.clone())
+            .route(&format!("{path}/"), method_router)
+    }
+}
