+++ tests/build.sh
@@ -1,13 +1,12 @@
 # build the docker images needed for the tests
-docker build ../modules/lfs-server -t "gaalafis/lfs-server:0.0.1-proxy-sbs" -f ../modules/lfs-server/Dockerfile --target=runtime_sbs_custom_signing
-docker build ../modules/lfs-server -t "gaalafis/lfs-server:0.0.1-proxy-sbs-locks-pg" -f ../modules/lfs-server/Dockerfile --target=runtime_sbs_custom_signing_locks
-docker build ../modules/lfs-server -t "gaalafis/lfs-server:0.0.1-signer-sbs-locks" -f ../modules/lfs-server/Dockerfile --target=runtime_main
+docker build ../modules/lfs-server -t "gaalafis/lfs-server:0.0.1-proxy-sbs" -f ../modules/lfs-server/Dockerfile --target=runtime-proxy-sbs
+docker build ../modules/lfs-server -t "gaalafis/lfs-server:0.0.1-proxy-sbs-locks-pg" -f ../modules/lfs-server/Dockerfile --target=runtime-proxy-sbs-locks-pg
+docker build ../modules/lfs-server -t "gaalafis/lfs-server:0.0.1-signer-sbs" -f ../modules/lfs-server/Dockerfile --target=runtime-signer-sbs
+docker build ../modules/lfs-server -t "gaalafis/lfs-server:0.0.1-signer-sbs-locks-pg" -f ../modules/lfs-server/Dockerfile --target=runtime-signer-sbs-locks-pg
+
 docker build ../modules/auth -t "gaalafis/gitolite:0.0.1" -f ../modules/auth/Dockerfile 
+
 docker build . -t "gaalafis:tester_client" -f ./runner/Dockerfile 
+
 docker build ./architectures/nginx -t "gaalafis:nginx" -f ./architectures/nginx/Dockerfile
 
-
-
-docker build ../modules/lfs-server -t "gaalafis/lfs-server:0.0.1-proxy-sbs" -f ../modules/lfs-server/Dockerfile --target=runtime_sbs_custom_signing
-
-docker build ./modules/lfs-server -t "gaalafis/lfs-server:0.0.1-signer-sbs" -f ./modules/lfs-server/Dockerfile.prod --build-arg binary=signer-sbs
